+++ test/cypress/e2e/with_fhir_server.spec.js
@@ -294,6 +294,33 @@ describe('FHIR Questionnaire Viewer with a specified FHIR server: ', () => {
           .should('not.be.visible');
     });
 
+    it('should support s=default', ()=> {
+      // For some reason this fails with version 29 of lforms, so test with
+      // lfv=latest.
+
+      // First make sure the server's version is used if the package is not
+      // present.
+      const serverBase = 'https://tx.fhir.org/r4'; // current default, though that might change
+      const appBase = Cypress.config().baseUrl;
+      const testDataBase = appBase;
+      const url = appBase + '/?lfv=latest&q=' + testDataBase +
+        '/questionaire-answerValueSet.json' + '&s=default';
+
+      cy.mockApiResponse(serverBase +
+        '/ValueSet/%24expand?url=http%3A%2F%2Fhl7.org%2Ffhir%2FValueSet%2Flanguage-preference-type*',
+        'ValueSet-preferred-language-type.json', 'serverVS');
+
+      cy.visit(url);
+      cy.wait('@serverVS');
+      const questionID = '1/1';
+      // Confirm that we are getting the (mocked) server copy.
+      cy.byId(questionID)
+          .type('{downArrow}')
+          .type('{enter}');
+      cy.byId(questionID)
+          .should('have.value', 'apples');
+    });
+
     it('should prefer to get ValueSets from the package rather than the FHIR' +
        'server if both are specified', ()=> {
       // For some reason this fails with version 29 of lforms, so test with